+++ sentry/src/test/java/io/sentry/SentryFeedbackOptionsTest.kt
@@ -1,6 +1,7 @@
 package io.sentry
 
 import io.sentry.SentryFeedbackOptions.IFormHandler
+import io.sentry.util.LoadClass
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import org.mockito.kotlin.mock
@@ -8,13 +9,14 @@ import org.mockito.kotlin.mock
 class SentryFeedbackOptionsTest {
   @Test
   fun `feedback options is initialized with default values`() {
-    val options = SentryFeedbackOptions(mock<IFormHandler>())
+    val options = SentryFeedbackOptions(mock<IFormHandler>(), LoadClass())
     assertEquals(false, options.isNameRequired)
     assertEquals(true, options.isShowName)
     assertEquals(false, options.isEmailRequired)
     assertEquals(true, options.isShowEmail)
     assertEquals(true, options.isUseSentryUser)
     assertEquals(true, options.isShowBranding)
+    assertEquals(true, options.isEnableAttachScreenshot)
     assertEquals("Report a Bug", options.formTitle)
     assertEquals("Send Bug Report", options.submitButtonLabel)
     assertEquals("Cancel", options.cancelButtonLabel)
@@ -26,6 +28,9 @@ class SentryFeedbackOptionsTest {
     assertEquals("Description", options.messageLabel)
     assertEquals("What's the bug? What did you expect?", options.messagePlaceholder)
     assertEquals("Thank you for your report!", options.successMessageText)
+    assertEquals("Add a screenshot", options.addScreenshotButtonLabel)
+    assertEquals("Remove screenshot", options.removeScreenshotButtonLabel)
+    assertEquals("Screenshot is too large", options.screenshotTooLargeMessageText)
     assertEquals(null, options.onFormOpen)
     assertEquals(null, options.onFormClose)
     assertEquals(null, options.onSubmitSuccess)
@@ -35,13 +40,14 @@ class SentryFeedbackOptionsTest {
   @Test
   fun `feedback options copy constructor`() {
     val options =
-      SentryFeedbackOptions(mock<IFormHandler>()).apply {
+      SentryFeedbackOptions(mock<IFormHandler>(), LoadClass()).apply {
         isNameRequired = true
         isShowName = false
         isEmailRequired = true
         isShowEmail = false
         isUseSentryUser = false
         isShowBranding = false
+        isEnableAttachScreenshot = false
         formTitle = "Title"
         submitButtonLabel = "Submit"
         cancelButtonLabel = "Cancel Label"
@@ -53,6 +59,9 @@ class SentryFeedbackOptionsTest {
         messageLabel = "Message Label"
         messagePlaceholder = "Message Placeholder"
         successMessageText = "Success Message"
+        addScreenshotButtonLabel = "Add Screenshot Label"
+        removeScreenshotButtonLabel = "Remove Screenshot Label"
+        screenshotTooLargeMessageText = "Too Large Message"
         onFormOpen = mock()
         onFormClose = mock()
         onSubmitSuccess = mock()
@@ -65,6 +74,7 @@ class SentryFeedbackOptionsTest {
     assertEquals(false, optionsCopy.isShowEmail)
     assertEquals(false, optionsCopy.isUseSentryUser)
     assertEquals(false, optionsCopy.isShowBranding)
+    assertEquals(false, optionsCopy.isEnableAttachScreenshot)
     assertEquals("Title", optionsCopy.formTitle)
     assertEquals("Submit", optionsCopy.submitButtonLabel)
     assertEquals("Cancel Label", optionsCopy.cancelButtonLabel)
@@ -76,6 +86,9 @@ class SentryFeedbackOptionsTest {
     assertEquals("Message Label", optionsCopy.messageLabel)
     assertEquals("Message Placeholder", optionsCopy.messagePlaceholder)
     assertEquals("Success Message", optionsCopy.successMessageText)
+    assertEquals("Add Screenshot Label", optionsCopy.addScreenshotButtonLabel)
+    assertEquals("Remove Screenshot Label", optionsCopy.removeScreenshotButtonLabel)
+    assertEquals("Too Large Message", optionsCopy.screenshotTooLargeMessageText)
     assertEquals(options.onFormOpen, optionsCopy.onFormOpen)
     assertEquals(options.onFormClose, optionsCopy.onFormClose)
     assertEquals(options.onSubmitSuccess, optionsCopy.onSubmitSuccess)